+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- Pipeline Settings: Allow to configure commands to run before deployment in pipeline configuration 
+- Pipeline Settings: Allow to configure commands to run before deployment in pipeline configuration
 
 ## [6.10.2] 2024-10-11